Исправил логику модулей
